moved component (Stack) around and extracted SimpleCard component
