+++ src/frontend/src/components/Forms/index.ts
@@ -1,4 +1,3 @@
 import FormContainer from './FormContainer';
-import Stack from './Stack';
 
-export { FormContainer, Stack };
+export { FormContainer };
