+++ src/frontend/src/components/index.tsx
@@ -1,4 +1,6 @@
 import VerticalDivider from './VerticalDivider/VerticalDivider';
 import GoToButton from './Buttons/GoTo';
+import SimpleCard from './SimpleCard/SimpleCard';
+import Stack from './Stack/Stack';
 
-export { GoToButton, VerticalDivider };
+export { Stack, GoToButton, VerticalDivider, SimpleCard };
