+++ src/frontend/src/pages/MyGroup/ProjectPreferenceCard.jsx
@@ -1,23 +1,14 @@
-import { Button } from '@carbon/react';
-import { ArrowRight } from '@carbon/icons-react';
-import { useNavigate } from 'react-router-dom';
 import AppRoutes from 'navigation/AppRoutes';
+import GoToButton from 'components/Buttons/GoTo';
 
 export function ProjectPreferenceCard({ title, rank, projectId }) {
-    const navigate = useNavigate();
-
     return (
         <div className='simple-card project'>
             <div>
                 <h5>Preference: #{rank}</h5>
                 {title}
             </div>
-            <Button
-                onClick={() => navigate(`${AppRoutes.project}/${projectId}`)}
-                renderIcon={ArrowRight}
-            >
-                Go to
-            </Button>
+            <GoToButton url={`${AppRoutes.project}/${projectId}`} />
         </div>
     );
 }
